fixed not being able to input JWT token in swagger and moved files to own package
