+++ src/main/java/no/ntnu/learniverseconnect/security/swagger/OpenApiConfig.java
@@ -1,0 +1,16 @@
+package no.ntnu.learniverseconnect.security.swagger;
+
+import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
+import io.swagger.v3.oas.annotations.security.SecurityScheme;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@SecurityScheme(
+    name = "bearerAuth",
+    type = SecuritySchemeType.HTTP,
+    scheme = "bearer",
+    bearerFormat = "JWT"
+)
+public class OpenApiConfig {
+  // No need for @OpenAPIDefinition or @Bean if you're adding security manually to each endpoint
+}
